refactor: move constants to root
