+++ cyberdrop_dl/clients/scraper_client.py
@@ -18,7 +18,7 @@ from aiohttp_client_cache.response import AnyResponse
 from aiohttp_client_cache.session import CachedSession
 from bs4 import BeautifulSoup
 
-import cyberdrop_dl.utils.constants as constants
+import cyberdrop_dl.constants as constants
 from cyberdrop_dl import env
 from cyberdrop_dl.exceptions import DDOSGuardError, DownloadError, InvalidContentTypeError, ScrapeError
 from cyberdrop_dl.utils.logger import log_debug
