+++ cyberdrop_dl/director.py
@@ -16,13 +16,12 @@ import browser_cookie3
 from pydantic import ValidationError
 from rich import print as rich_print
 
-from cyberdrop_dl import env
+from cyberdrop_dl import constants, env
 from cyberdrop_dl.exceptions import InvalidYamlError
 from cyberdrop_dl.managers.manager import Manager
 from cyberdrop_dl.scraper.scrape_mapper import ScrapeMapper
 from cyberdrop_dl.ui.program_ui import ProgramUI
 from cyberdrop_dl.ui.textual import textual_ui
-from cyberdrop_dl.utils import constants
 from cyberdrop_dl.utils.apprise import send_apprise_notifications
 from cyberdrop_dl.utils.dumper import Dumper
 from cyberdrop_dl.utils.logger import LogHandler, QueuedLogger, TextualLogQueueHandler, log, log_spacer, log_with_color
