+++ cyberdrop_dl/utils/yaml.py
@@ -11,8 +11,8 @@ import yaml
 from pydantic import BaseModel, ValidationError
 from yarl import URL
 
+from cyberdrop_dl.constants import CLI_VALIDATION_ERROR_FOOTER, VALIDATION_ERROR_FOOTER
 from cyberdrop_dl.exceptions import InvalidYamlError
-from cyberdrop_dl.utils.constants import CLI_VALIDATION_ERROR_FOOTER, VALIDATION_ERROR_FOOTER
 
 if TYPE_CHECKING:
     from pydantic_core import ErrorDetails
